init: do not execute `gitlab:check` rake task
In the forthcoming commit all daemons will be auto started using
supervisord and the execution of the init script will end at the launch
of supervisord. This does not allow room for the `gitlab:check` task to
be executed in such a way that its output.

User's will need to login to the container and manually start the
`gitlab:check` task if the wish to check the status.