Add configuration for sendmail
If `SENDMAIL_ENABLED` is set to `true` then postfix will be started
and Gitlab will be configured to use sendmail. If both
`SENDMAIL_ENABLED` and `SMTP_ENABLED` are `true` then the sendmail
feature is disabled. The configuration for postfix is simple; it is
setup as a relay.

Postfix forks into the background and as a result cannot be started
under supervisord. See the following link for more background
information:

http://goo.gl/F4neAQ

Closes sameersbn/docker-gitlab#208