Create app:init to separate initialization from starting
This feature allows an install script to initialize much of the
gitlab requirements without actually starting it.  It is useful
for deployment scripts so that the first start up does not have to
do this work itself.  app:start calls app:init, so no change
in functionality for app:start.