Implement unix socket connection to redis
This is a first step to create a full unix socket based scheme of
communications between the related containers.

Sockets-dvc is a new DVC that will contain all sockets.

To control how the connection to the redis container is done we add a
new configuration option REDIS_CONNECTION, which can be either "tcp" or
"unix"