Implement unix socket connection to redis
This is the first step to create a full unix socket based scheme of
communications between the related containers.

Sockets-dvc is a new DVC that will contain all sockets.

The socket connection is used if the linked redis container is used and
a socket is available inside the docker container. Otherwise, we fall
back to a tcp connection.