Segregate nginx to a separate container
Following isolation principles of docker the nginx process is moved to a separate
container, instead running in the gitlab container. The nginx container uses an
official nginx image. The configuration scripts detects if an nginx container is
linked and does not start nginx inside the gitlab container in that case. If no
nginx container is detected the script works as normal.

Important changes:
* The gitlab-workhorse socket is moved to a more reasonable location
* Variable `NGINX_USE_EXTERNAL`. If true, the gitlab container does not start
  an nginx process, even if an nginx container is not detected. Useful if one have
  an nginx process outside of docker-compose