Downgrade to git 2.10.2.
Due to issue
[995](https://github.com/sameersbn/docker-gitlab/issues/995)
which in turn is due to issue
[25301](https://gitlab.com/gitlab-org/gitlab-ce/issues/25301)
git version 2.11.0 (a recent ubuntu 14.04 upgrade) is not working
correctly.  This removes the git-core ubuntu package install
and replaces it with a git install from source. It does so
in an easily removable block.

Note that the image will be a bit larger than normal.  The
built git needs some libraries installed and I didn't spend
the time figuring out which.