clean up the $HOME/.cache directory after installation
Yarn stores every package in a global cache in your user directory on
the file system. This ends up to be around ~400M of extra utilized by
the image and has no benefits whatsoever.

Additionally `go` populates a build cache with will also be cleared
when the ~/.cache directory is purged

/cc @solidnerd