Improvements to readme.md : database restore, mysql, pg_trgm, & app:rake
## Fixed the restore command 

The restore command was incorrect, the time stamp being used is no longer valid.

## Added another reference on how to interact with app:rake

For debugging purposes it can be immensely helpful to perform a series of commands on an interactive terminal.

This command outlines how users can interact with app:rake commands when attached to the container.

## Describe how to add `pg_trgm` extension to postgres

The `pg_trgm` extension has been required for quite a few versions. The specific SQL command to add this extension was been missing. This commit adds the SQL command to create that extension.

## MySQL is deprecated

A brief statement had mentioned that Gitlab HQ discourages using MySQL. It did not offer any insight, or reference as to why, however. I included a little bit more detail and a link to Gitlab HQ's official statement.

Closes sameersbn/docker-gitlab#1336