Use updated gitlab-ce repository name `gitlab-foss`
Gitlab is moving to a single codebase for GitLab Community and Enterprise editions.
All code and development is moved into the gitlab-ee repository.
The code of the gitlab-ce repository remains as is, and won't include any proprietary code.
The `gitlab-ce` repository is renamed to `gitlab-foss`

See https://about.gitlab.com/2019/08/23/a-single-codebase-for-gitlab-community-and-enterprise-edition/

Signed-off-by: Pierre Belloy <p.belloy@axelor.com>